Add additional checks to various functions

+++ handlers/deletecomment.py
@@ -1,5 +1,7 @@
 from handler import Handler
 from models import Comments
+from util import comment_exists
+
 
 class DeleteComment(Handler):
     def get(self):
@@ -7,10 +9,14 @@ class DeleteComment(Handler):
 
     def post(self):
         comment_id = int(self.request.get("cid"))
+        blog_id = int(self.request.get("bid"))
 
-        comment = Comments.by_id(comment_id)
-        blog_id = comment.blog_id
-        if comment and comment.username == self.user:
-            comment.delete()
+        # Checks user's auth status and comment exists
+        if self.user and comment_exists(comment_id):
+            comment = Comments.by_id(comment_id)
+
+            # Checks to see if user owns the comment
+            if comment and comment.username == self.user:
+                comment.delete()
 
         self.redirect("/post/%s" % blog_id)
